Fixing tests
