+++ reflection-config-generator/src/main/java/nativeimage/core/ReflectionUtils.java
@@ -1,5 +1,9 @@
 package nativeimage.core;
 
+import nativeimage.Reflection;
+
+import org.apache.commons.lang3.StringUtils;
+
 import javax.lang.model.type.MirroredTypeException;
 
 public class ReflectionUtils {
@@ -10,4 +14,20 @@ public class ReflectionUtils {
 			return e.getTypeMirror().toString().startsWith(Void.class.getName());
 		}
 	}
+
+	public static String getClassName(Reflection reflectionAnn){
+		return StringUtils.firstNonBlank(getScanClass(reflectionAnn), reflectionAnn.scanClassName());
+	}
+
+	private static String getScanClass(Reflection reflectionAnn) {
+		if (ReflectionUtils.isVoid(reflectionAnn)) {
+			return null;
+		}
+		try {
+			return reflectionAnn.scanClass().getName();
+		} catch (MirroredTypeException e) {
+			return e.getTypeMirror().toString();
+		}
+	}
+
 }
