+++ reflection-config-generator/src/main/java/nativeimage/core/TypeBuilder.java
@@ -3,8 +3,6 @@ package nativeimage.core;
 import java.util.Collections;
 import java.util.Set;
 
-import javax.lang.model.type.MirroredTypeException;
-
 import com.mageddo.aptools.ClassUtils;
 
 import org.apache.commons.lang3.StringUtils;
@@ -18,9 +16,7 @@ public final class TypeBuilder {
 	}
 
 	public static Set<String> of(Reflection reflectionAnn, String clazzName) {
-		final String expectedClassName = StringUtils.firstNonBlank(
-				getScanClass(reflectionAnn), reflectionAnn.scanClassName()
-		);
+		final String expectedClassName = ReflectionUtils.getClassName(reflectionAnn);
 		if(StringUtils.isNotBlank(expectedClassName)){
 			if(ClassUtils.doClassOwnPossibleSubClassOrIsTheSame(expectedClassName, clazzName)){
 				return toSet(clazzName);
@@ -33,17 +29,6 @@ public final class TypeBuilder {
 			return Collections.emptySet();
 		}
 		return toSet(clazzName);
-	}
-
-	private static String getScanClass(Reflection reflectionAnn) {
-		if (reflectionAnn.scanClass() == Void.class) {
-			return null;
-		}
-		try {
-			return reflectionAnn.scanClass().getName();
-		} catch (MirroredTypeException e) {
-			return e.getTypeMirror().toString();
-		}
 	}
 
 	private static Set<String> toSet(String type) {
